% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_regression_model.R
\name{fit_regression_model}
\alias{fit_regression_model}
\title{Fit regression model.}
\usage{
fit_regression_model(dat, type = "linear")
}
\arguments{
\item{dat}{is a dataframe}

\item{type}{is a string indicating linear or polynomial
regression to be fitted.}
}
\value{
it returns a regression model
}
\description{
This function fits a regression model to the df dat. The type of regression
fit is modified by the variable type.
}
